requirements.txt
numpy==2.0.0
opencv-python==4.10.0.84
pillow==10.4.0
streamlit==1.36.0
